Move roles functions to roles.functions and target es2016
