+++ src/app/functions/index.ts
@@ -1,3 +1,4 @@
 export * from './capitalize.functions';
 export * from './date-format.functions';
 export * from './helpers.functions';
+export * from './roles.functions';
