+++ src/app/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es6",
+    "target": "es2016",
     "sourceMap": true,
     "experimentalDecorators": true
   },
